Decouple db from web server
This commit decouples the db instance from
the mediawiki web server to allow instances to
use a remote mysql db.

- adds a parameter db_server to mediawiki::instance
that can be used to specify remote dbs.
- adds a parameter db_install that can be used
to specify if a db should be installed.
- adds the class mediawiki::db::access that can
be used to allow remote privs for wikimedia instances
to remotely create users and tables.