Permit toggling of default Apache vhost creation
In the event of running a MediaWiki instance on a nonstandard port and a
reverse proxy on port 80, both on the same host, you need to not have
Apache listening on port 80.  Unfortunately, simply passing the
default_vhost parameter causes a dependency cycle :( so there are some
hackish fixes for that problem.